+++ Minigin/Transform.h
@@ -6,8 +6,8 @@ namespace GameEngine
 	class Transform final
 	{
 	public:
-		[[nodiscard]] glm::ivec3 GetPosition() const { return m_Position; }
-		void SetPosition(int x, int y, int z);
+		[[nodiscard]] glm::vec3 GetPosition() const { return m_Position; }
+		void SetPosition(float x, float y, float z);
 		void Translate(int x, int y);
 		void Translate(const glm::ivec2& vec);
 		void SetRotation(float angle);
